pyproject.toml
[project]
name = "receipt-transcription-sim"
version = "1.0.0"
description = "Discrete-event simulation of a receipt transcription service"
requires-python = ">=3.11"
dependencies = [
    "numpy>=1.24",
    "pyyaml>=6.0",
    "pandas>=2.0",
]

[project.scripts]
receipt-sim = "receipt_sim.cli:main"

[dependency-groups]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "scipy>=1.10",
    "mypy>=1.0",
    "flake8>=7.0",
    "black>=24.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.black]
line-length = 88
target-version = ["py311"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/receipt_sim"]
